+++ library/Episciences/VolumesManager.php
@@ -3,6 +3,7 @@
 class Episciences_VolumesManager
 {
     public const MAX_STRING_LENGTH = 255;
+    public const MAX_STRING_LENGTH_VOL_NUM = 6;
 
     /**
      * Retourne un volume
@@ -350,8 +351,20 @@ class Episciences_VolumesManager
                 [new Zend_Validate_StringLength(['max' => self::MAX_STRING_LENGTH])]
             ],
         ]);
+        $form->addElement('text', 'num', [
+            'label' => 'Numéro du volume',
+            'placeholder' => Zend_Registry::get('Zend_Translate')->translate('Numéro du volume'),
+            'value' => ($volume !== null) ? $volume->getVol_num() : "",
+            'required' => true,
+            'style' => 'width:300px;position: static;',
+            'validators' => [
+                [new Zend_Validate_StringLength(['max' => self::MAX_STRING_LENGTH_VOL_NUM])],
+            ],
+        ]);
+
         $form->addElement('text', 'year', [
             'label' => 'Année du volume',
+            'value' => ($volume !== null) ? $volume->getVol_year() : '',
             'required' => true,
             'style' => 'width:300px;position: static;',
             'validators' => [